pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "inboxagent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "msal>=1.31.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "aiohttp>=3.10.0",
    "pydantic-settings>=2.5.0",
    "pytz>=2024.2",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
]

[project.scripts]
inboxagent = "inboxagent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
